+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.2.0 (Next)
 **FaaS**
 - Return `Boolean` for `validateTemplate`.
+- Add `list` method.
 
 **GoSS**
 - Increase minimum supported version to 0.4.